refactor(xml): type the status and node-type codes
Replace the bare i32/u32 status and node-type constants in xml/abi.rs with
NodeType, Status and MutStatus enums, and thread them through the arena, tree
builder, mutators, ffi boundary, glue, cross-import and the XPath XML backend.
Discriminants are unchanged, so the values that cross to Ruby (#node_type) and
the exception classes stay identical.

- NodeType is the WHATWG DOM nodeType numbering (TryFrom<u32> rejects the
  unused 5/6/12); Node.type_ is a NodeType and Document::type_ returns
  Option<NodeType>, an invalid handle mapping to None.
- Mutation failures get their own MutStatus instead of overloading the parse
  Status. The null/stale-document case is now MutStatus::Internal (9) rather
  than the parse code 4, which mkr_xml_mut_check had been reporting as
  "namespace prefix is not bound". mkr_xml_mut_check is exhaustive (the
  wildcard fallback is gone) and is a plain fn, not extern "C" (no C caller
  remains).
- remove_attribute/_ns return bool; Document.has_encoding_decl is bool; the
  deep / inherit_doc_ns flags are bool rather than i32 (matching clone_node).

Gates: cargo clippy -D warnings (both feature sets), cargo fmt --check, cargo
test --no-default-features (9/0), rake spec (1006/0), rake symbols, fuzz:mutate
and fuzz:xml (no findings).

+++ ext/makiri/rust/src/glue/doc.rs
@@ -390,13 +390,7 @@ fn doc_import_node(ruby: &Ruby, self_: Value, args: &[Value]) -> Result<Value, E
             );
             let src =
                 crate::xml::abi::NodeId::from_token(mkr_xml_node_unwrap(node_v.as_raw()) as usize);
-            mkr_xml_mut_check(mkr_cross_xml_to_html(
-                doc,
-                xdoc,
-                src,
-                i32::from(deep),
-                &mut imp,
-            ));
+            mkr_xml_mut_check(mkr_cross_xml_to_html(doc, xdoc, src, deep, &mut imp));
             return Ok(Value::from_raw(mkr_wrap_html_node(imp, self_.as_raw())));
         }
